client/requirements.txt
diffiehellman
pycrypto